# internal/handler/gym_handler.go
package handler

import (
	"net/http"
	"strconv"

	"Gym_StrongCode/internal/service"

	"github.com/gin-gonic/gin"
)

type GymHandler struct {
	gymService *service.GymService
}

func NewGymHandler(gymService *service.GymService) *GymHandler {
	return &GymHandler{gymService: gymService}
}

// ListGyms godoc
// @Summary      List all gyms
// @Description  Get all gym locations
// @Tags         gyms
// @Produce      json
// @Success      200  {array}   models.Gym
// @Failure      500  {object}  map[string]string
// @Router       /gyms [get]
func (h *GymHandler) List(c *gin.Context) {
	gyms, err := h.gymService.List()
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gyms)
}

type createGymRequest struct {
	Name    string `json:"name" binding:"required"`
	Address string `json:"address"`
}

// CreateGym godoc
// @Summary      Create gym
// @Description  Create a new gym location (admin only)
// @Tags         gyms
// @Security     Bearer
// @Accept       json
// @Produce      json
// @Param        body  body      handler.createGymRequest  true  "Gym data"
// @Success      201   {object}  models.Gym
// @Failure      400   {object}  map[string]string
// @Router       /admin/gyms [post]
func (h *GymHandler) Create(c *gin.Context) {
	var req createGymRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	gym, err := h.gymService.Create(req.Name, req.Address)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, gym)
}

type updateGymRequest struct {
	Name    string `json:"name" binding:"required"`
	Address string `json:"address"`
}

// UpdateGym godoc
// @Summary      Update gym
// @Description  Update gym details (admin only)
// @Tags         gyms
// @Security     Bearer
// @Accept       json
// @Produce      json
// @Param        id    path      int                      true  "Gym ID"
// @Param        body  body      handler.updateGymRequest true  "Updated gym data"
// @Success      200   {object}  map[string]string
// @Failure      400   {object}  map[string]string
// @Router       /admin/gyms/{id} [put]
func (h *GymHandler) Update(c *gin.Context) {
	idStr := c.Param("id")
	id, _ := strconv.Atoi(idStr)

	var req updateGymRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if err := h.gymService.Update(id, req.Name, req.Address); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"message": "gym updated"})
}

// DeleteGym godoc
// @Summary      Delete gym
// @Description  Delete gym location (admin only)
// @Tags         gyms
// @Security     Bearer
// @Param        id   path      int  true  "Gym ID"
// @Success      200  {object}  map[string]string
// @Failure      400  {object}  map[string]string
// @Router       /admin/gyms/{id} [delete]
func (h *GymHandler) Delete(c *gin.Context) {
	idStr := c.Param("id")
	id, _ := strconv.Atoi(idStr)

	if err := h.gymService.Delete(id); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"message": "gym deleted"})
}